HOUSE RENTAL MANAGEMENT APP
This app is a house rental management app. This app is basically used to rent out your apartment and houses to other people. In this app, the house owner will place their houses on the app and will add the details of the houses. The user will who wants houses on rent will see the house on the app and contact the owner for further procedures.